requirements.txt
numpy
matplotlib
pygame
Pillow